all: add c & rust compilation options

+++ upatch-diff/upatch-elf.h
@@ -87,7 +87,7 @@ struct rela {
 	GElf_Rela rela;
 	struct symbol *sym;
 	unsigned int type;
-	unsigned int offset;
+	unsigned long offset;
 	long addend;
 	char *string;
 	bool need_dynrela;
